Update main.yml
Optimizing the workflow🔗  Next, let’s look at how we can optimize the GitHub Actions workflow through build cache. This has two main advantages:

Build cache reduces the build time as it will not have to re-download all of the images, and
It also reduces the number of pulls we complete against Docker Hub. We need to make use of GitHub cache to make use of this.